requirements.txt
streamlit
rembg
Pillow
onnxruntime 